task 5

+++ bit_manipulation/5-flip_bits.c
@@ -7,8 +7,8 @@
 */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned long int cont = 0;
-	int i;
+	int cont = 0;
+	unsigned long int i;
 
 	for (i = n ^ m; i != 0; i = i & (i - 1))
 		cont++;
